1) Improved delete entities system. 2) findAll's total annihilation (where it was needed). 3) Fixed car's validation; 4) Abstract things became more abstract. 5) Wraps added. 6) Consultant's rate fixed. 7) Builders for DTO's added. 8) General and cosmetic improves implemented. 9) Fresh JSON for Postman and DDL+DML added.

+++ src/main/java/carDealerApplication/service/impl/AdministrationServiceImpl.java
@@ -3,6 +3,7 @@ package carDealerApplication.service.impl;
 import carDealerApplication.api.dto.AdministratorDTO;
 import carDealerApplication.dal.AdministratorRepository;
 import carDealerApplication.entity.Administrator;
+import carDealerApplication.exception.EntityNotFoundException;
 import carDealerApplication.exception.UserNotFoundException;
 import carDealerApplication.service.AdministratorService;
 import carDealerApplication.service.dtoConverter.AdministratorDTOConverter;
@@ -34,7 +35,12 @@ public class AdministrationServiceImpl implements AdministratorService {
 
     @Override
     public void deleteEntityById(Long administratorId) {
-        administratorRepository.deleteById(administratorId);
+        if (administratorRepository.existsById(administratorId)) {
+            administratorRepository.deleteById(administratorId);
+        } else {
+            throw new EntityNotFoundException("User not found.");
+        }
+
     }
 
     @Override
