+++ src/main/java/carDealerApplication/service/impl/CarServiceImpl.java
@@ -3,12 +3,14 @@ package carDealerApplication.service.impl;
 import carDealerApplication.api.dto.CarDTO;
 import carDealerApplication.dal.CarRepository;
 import carDealerApplication.entity.Car;
+import carDealerApplication.exception.EntityNotFoundException;
 import carDealerApplication.service.CarService;
 import carDealerApplication.service.dtoConverter.CarDTOConverter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 @Service
@@ -24,47 +26,33 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public List<Car> fetchEntityList() {
-        List<Car> carList = new ArrayList<>();
-        for (Car car : carRepository.findAll()) {
-            carList.add(car);
-        }
-        return carList;
+        return new ArrayList<>((Collection<? extends Car>) carRepository.findAll());
     }
 
     @Override
     public void deleteEntityById(Long carId) {
-        carRepository.deleteById(carId);
+        if(carRepository.existsById(carId)){
+            carRepository.deleteById(carId);
+        } else {
+            throw new EntityNotFoundException("Car not found.");
+        }
     }
 
     @Override
     public List<Car> fetchAvailableCar() {
-        List<Car> carList = new ArrayList<>();
-        for (Car car : carRepository.findAll()) {
-            if (car.isAvailable()) {
-                carList.add(car);
-            }
-        }
-        return carList;
+        return new ArrayList<>(carRepository.findCarsByIsAvailable(true));
     }
 
     @Override
     public List<Car> fetchUnavailableCar() {
-        List<Car> carList = new ArrayList<>();
-        for (Car car : carRepository.findAll()) {
-            if (!car.isAvailable()) {
-                carList.add(car);
-            }
-        }
-        return carList;
+        return new ArrayList<>(carRepository.findCarsByIsAvailable(false));
     }
 
     @Override
-    public List<CarDTO> optimizedFetchAvailableCar() {
+    public List<CarDTO> limitedFetchAvailableCar() {
         List<CarDTO> carDTOList = new ArrayList<>();
-        for (Car car : carRepository.findAll()) {
-            if (car.isAvailable()) {
-                carDTOList.add(carDTOConverter.convertToDTO(car));
-            }
+        for (Car car : carRepository.findCarsByIsAvailable(true)) {
+            carDTOList.add(carDTOConverter.convertToDTO(car));
         }
         return carDTOList;
     }
