+++ src/main/java/carDealerApplication/service/impl/SpecialOfferServiceImpl.java
@@ -1,19 +1,22 @@
 package carDealerApplication.service.impl;
 
+import carDealerApplication.dal.LocationRepository;
 import carDealerApplication.dal.SpecialOfferRepository;
-import carDealerApplication.entity.Location;
 import carDealerApplication.entity.SpecialOffer;
+import carDealerApplication.exception.EntityNotFoundException;
 import carDealerApplication.service.SpecialOfferService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class SpecialOfferServiceImpl implements SpecialOfferService {
     private final SpecialOfferRepository specialOfferRepository;
+    private final LocationRepository locationRepository;
 
     @Override
     public void upsertEntity(SpecialOffer specialOffer) {
@@ -22,28 +25,21 @@ public class SpecialOfferServiceImpl implements SpecialOfferService {
 
     @Override
     public List<SpecialOffer> fetchEntityList() {
-        List<SpecialOffer> specialOfferList = new ArrayList<>();
-        for (SpecialOffer specialOffer : specialOfferRepository.findAll()) {
-            specialOfferList.add(specialOffer);
-        }
-        return specialOfferList;
+        return new ArrayList<>((Collection<? extends SpecialOffer>) specialOfferRepository.findAll());
     }
 
     @Override
     public void deleteEntityById(Long specialOfferId) {
-        specialOfferRepository.deleteById(specialOfferId);
+        if (specialOfferRepository.existsById(specialOfferId)) {
+            specialOfferRepository.deleteById(specialOfferId);
+        } else {
+            throw new EntityNotFoundException("Special offer not found.");
+        }
     }
 
     @Override
     public List<SpecialOffer> fetchSpecialOfferByCountry(String country) {
-        List<SpecialOffer> specialOfferList = new ArrayList<>();
-        for (SpecialOffer specialOffer : specialOfferRepository.findAll()) {
-            for (Location location : specialOffer.getCountries()) {
-                if (country.equals(location.getCountry())) {
-                    specialOfferList.add(specialOffer);
-                }
-            }
-        }
-        return specialOfferList;
+        return new ArrayList<>(specialOfferRepository.findSpecialOfferByCountriesIn(locationRepository.getLocationsByCountry(country)));
     }
+
 }
